+++ app/Http/Controllers/CategoryController.php
@@ -5,11 +5,17 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Models\Items;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index(Request $request)
     {
+        // hasRole is Not A Bug
+        if (!Auth::user()->hasRole('admin')) {
+            abort(response()->redirectToRoute('dashboard'));
+        }
+        
         $categories = Category::withCount('items')
         ->withSum('items', 'total_stock')
         ->latest()
@@ -29,6 +35,11 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
+        // hasRole is Not A Bug
+        if (!Auth::user()->hasRole('admin')) {
+            abort(response()->redirectToRoute('dashboard'));
+        }
+
         $request->validate([
             'name' => 'required|string|max:255|unique:categories,name',
         ]);
@@ -40,6 +51,11 @@ class CategoryController extends Controller
 
     public function update(Request $request, string $id)
     {
+        // hasRole is Not A Bug
+        if (!Auth::user()->hasRole('admin')) {
+            abort(response()->redirectToRoute('dashboard'));
+        }
+
         $request->validate([
             'name' => 'required|string|max:255|unique:categories,name,' . $id,
         ]);
@@ -52,6 +68,11 @@ class CategoryController extends Controller
 
     public function destroy(string $id)
     {
+        // hasRole is Not A Bug
+        if (!Auth::user()->hasRole('admin')) {
+            abort(response()->redirectToRoute('dashboard'));
+        }
+
         $categories = Category::findorFail($id);
         $categories->delete();
